api/requirements.txt
flask
pillow
transformers
torch